fix : fix the plant-add
