+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './get-plant-detail';
 export * from './get-plant-list';
+export * from './upload-photo';
