+++ src/hooks/upload-photo/index.ts
@@ -1,0 +1,1 @@
+export * from './upload-photo.mutation';
