+++ src/hooks/upload-photo/upload-photo.mutation.ts
@@ -1,0 +1,52 @@
+import { CLOUDINARY_NAME } from '@env';
+import axios from 'axios';
+import isUndefined from 'lodash/isUndefined';
+import { launchCamera } from 'react-native-image-picker';
+import { useMutation } from 'react-query';
+
+export const useUploadPhotoMutation = () => {
+  const makePhotoFormData = async () => {
+    const result = await launchCamera({
+      mediaType: 'photo',
+      saveToPhotos: true,
+    });
+
+    if (isUndefined(result.assets) || result.didCancel) {
+      return;
+    }
+
+    const { uri, fileName, type } = result.assets[0];
+
+    const source = {
+      uri,
+      type,
+      name: fileName,
+    };
+    const formData = new FormData();
+    const cloudName = CLOUDINARY_NAME;
+
+    formData.append('file', source);
+    formData.append('upload_preset', 'BloomMate');
+    formData.append('cloud_name', cloudName);
+
+    return formData;
+  };
+
+  return useMutation(async () => {
+    const data = await makePhotoFormData();
+    const cloudName = CLOUDINARY_NAME;
+
+    const response = await axios(
+      `https://api.cloudinary.com/v1_1/${cloudName}/image/upload`,
+      {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'multipart/form-data',
+        },
+        data,
+      },
+    );
+
+    return response;
+  });
+};
