+++ src/screens/plant-add/hooks/reset-plant-state/reset-plant-state.hook.ts
@@ -1,0 +1,12 @@
+import { useResetRecoilState } from 'recoil';
+import { useWillUnmount } from 'rooks';
+
+import { $plantAddState } from '../../plant-add.state';
+
+export const useResetPlantState = () => {
+  const resetState = useResetRecoilState($plantAddState);
+
+  useWillUnmount(() => {
+    resetState();
+  });
+};
