--- src/screens/plant-add/modules/plant-add-content/components/plant-add-content-photo/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './upload-photo';
